requirements.txt
boto3>=1.34.11
dropbox>=11.36.2
ffmpeg-python>=0.2.0
google-cloud-aiplatform>=1.38.1
google-cloud-storage>=2.14.0
notion-client>=2.2.1